Rename: "패키지 이름 변경"
- domain이라는 클래스명에서 세부 분류를 위한 준비과정으로 model이라는 패키지명로 변경
