+++ src/main/java/baseball/model/GameCycle.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.model;
 
 public class GameCycle {
     private Result result;
