+++ src/main/java/baseball/model/NumbersValidator.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.model;
 
 public class NumbersValidator extends Validator{
     @Override
