+++ src/test/java/baseball/model/NumbersTest.java
@@ -1,6 +1,5 @@
 package baseball.model;
 
-import baseball.domain.Numbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
