+++ src/test/java/baseball/model/ResultTest.java
@@ -1,7 +1,5 @@
 package baseball.model;
 
-import baseball.domain.Numbers;
-import baseball.domain.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
